First implementation of JSON-based themes
Themes are loaded from the root "themes" folder, I placed a sample documentation txt file to describe the JSON structure.
Old image background was replaced with a dynamic background with multiple options.
UXML elements are dynamically recolored based on the loaded settings, still WIP to make it work everywhere in the game.
I made a few example themes ready to be loaded too.