Change SongSearch behaviour
The SongSearch currently searches for the whole string as given. This causes problems and confusions when using to many whitespaces ("ABBA " or "Imagine  Dragons" would not find anything) and it makes it difficult to search for artist and title at the same time ("Queen Bohemian" does not find anything). I have seen a lot of people confused about both of these problems.

The more intuitive behaviour would be to split the search string at all whitespaces and to search for each part individually. This small edit should create this behaviour (but I did not test it).